CVSL-618 add PrisonApiClient.kt webclient and latest HDC status Configure new webclient using client credentials for PrisonAPI. Use latest HDC status to avoid emailing sentence changes where the current licence has been superseded by a HDC approved licence.

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/service/prison/PrisonerHdcStatus.kt
@@ -3,7 +3,7 @@ package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.prison
 data class PrisonerHdcStatus(
   val approvalStatus: String?,
   val approvalStatusDate: String?,
-  val bookingId: Int?,
+  val bookingId: Long?,
   val checksPassedDate: String?,
   val passed: Boolean,
   val refusedReason: String?
